Joined openvino files in one bundle (#5)
* dot rendering is broken

* almost abstracted the node

* finished with anstracting the node
