+++ setup.py
@@ -119,8 +119,7 @@ setuptools.setup(
             'sklearn.js', 'sklearn-metadata.json',
             'tf.js', 'tf-metadata.json', 'tf-proto.js', 
             'tflite.js', 'tflite-metadata.json', 'tflite-schema.js',
-            'openvino-ir.js', 'openvino-ir-proto.js',
-            'openvino-dot.js', 'openvino-dot-proto.js',
+            'openvino-ir.js', 'openvino-ir-parser.js',
             'view-browser.html', 'view-browser.js',
             'view-grapher.css', 'view-grapher.js',
             'view-sidebar.css', 'view-sidebar.js',
