+++ src/openvino-ir-parser.js
@@ -3550,3 +3550,2542 @@ window.OpenVINOIRParser = /*
             parse:       peg$parse
         };
     })();
+
+    window.OpenVINODotParser = /*
+    * Generated by PEG.js 0.10.0.
+    *
+    * http://pegjs.org/
+    */
+       (function() {
+           "use strict";
+   
+           function peg$subclass(child, parent) {
+               function ctor() { this.constructor = child; }
+               ctor.prototype = parent.prototype;
+               child.prototype = new ctor();
+           }
+   
+           function peg$SyntaxError(message, expected, found, location) {
+               this.message  = message;
+               this.expected = expected;
+               this.found    = found;
+               this.location = location;
+               this.name     = "SyntaxError";
+   
+               if (typeof Error.captureStackTrace === "function") {
+                   Error.captureStackTrace(this, peg$SyntaxError);
+               }
+           }
+   
+           peg$subclass(peg$SyntaxError, Error);
+   
+           peg$SyntaxError.buildMessage = function(expected, found) {
+               var DESCRIBE_EXPECTATION_FNS = {
+                   literal: function(expectation) {
+                       return "\"" + literalEscape(expectation.text) + "\"";
+                   },
+   
+                   "class": function(expectation) {
+                       var escapedParts = "",
+                           i;
+   
+                       for (i = 0; i < expectation.parts.length; i++) {
+                           escapedParts += expectation.parts[i] instanceof Array
+                               ? classEscape(expectation.parts[i][0]) + "-" + classEscape(expectation.parts[i][1])
+                               : classEscape(expectation.parts[i]);
+                       }
+   
+                       return "[" + (expectation.inverted ? "^" : "") + escapedParts + "]";
+                   },
+   
+                   any: function(expectation) {
+                       return "any character";
+                   },
+   
+                   end: function(expectation) {
+                       return "end of input";
+                   },
+   
+                   other: function(expectation) {
+                       return expectation.description;
+                   }
+               };
+   
+               function hex(ch) {
+                   return ch.charCodeAt(0).toString(16).toUpperCase();
+               }
+   
+               function literalEscape(s) {
+                   return s
+                       .replace(/\\/g, '\\\\')
+                       .replace(/"/g,  '\\"')
+                       .replace(/\0/g, '\\0')
+                       .replace(/\t/g, '\\t')
+                       .replace(/\n/g, '\\n')
+                       .replace(/\r/g, '\\r')
+                       .replace(/[\x00-\x0F]/g,          function(ch) { return '\\x0' + hex(ch); })
+                       .replace(/[\x10-\x1F\x7F-\x9F]/g, function(ch) { return '\\x'  + hex(ch); });
+               }
+   
+               function classEscape(s) {
+                   return s
+                       .replace(/\\/g, '\\\\')
+                       .replace(/\]/g, '\\]')
+                       .replace(/\^/g, '\\^')
+                       .replace(/-/g,  '\\-')
+                       .replace(/\0/g, '\\0')
+                       .replace(/\t/g, '\\t')
+                       .replace(/\n/g, '\\n')
+                       .replace(/\r/g, '\\r')
+                       .replace(/[\x00-\x0F]/g,          function(ch) { return '\\x0' + hex(ch); })
+                       .replace(/[\x10-\x1F\x7F-\x9F]/g, function(ch) { return '\\x'  + hex(ch); });
+               }
+   
+               function describeExpectation(expectation) {
+                   return DESCRIBE_EXPECTATION_FNS[expectation.type](expectation);
+               }
+   
+               function describeExpected(expected) {
+                   var descriptions = new Array(expected.length),
+                       i, j;
+   
+                   for (i = 0; i < expected.length; i++) {
+                       descriptions[i] = describeExpectation(expected[i]);
+                   }
+   
+                   descriptions.sort();
+   
+                   if (descriptions.length > 0) {
+                       for (i = 1, j = 1; i < descriptions.length; i++) {
+                           if (descriptions[i - 1] !== descriptions[i]) {
+                               descriptions[j] = descriptions[i];
+                               j++;
+                           }
+                       }
+                       descriptions.length = j;
+                   }
+   
+                   switch (descriptions.length) {
+                       case 1:
+                           return descriptions[0];
+   
+                       case 2:
+                           return descriptions[0] + " or " + descriptions[1];
+   
+                       default:
+                           return descriptions.slice(0, -1).join(", ")
+                               + ", or "
+                               + descriptions[descriptions.length - 1];
+                   }
+               }
+   
+               function describeFound(found) {
+                   return found ? "\"" + literalEscape(found) + "\"" : "end of input";
+               }
+   
+               return "Expected " + describeExpected(expected) + " but " + describeFound(found) + " found.";
+           };
+   
+           function peg$parse(input, options) {
+               options = options !== void 0 ? options : {};
+   
+               var peg$FAILED = {},
+   
+                   peg$startRuleFunctions = { start: peg$parsestart },
+                   peg$startRuleFunction  = peg$parsestart,
+   
+                   peg$c0 = peg$otherExpectation("graph"),
+                   peg$c1 = "strict",
+                   peg$c2 = peg$literalExpectation("strict", true),
+                   peg$c3 = "graph",
+                   peg$c4 = peg$literalExpectation("graph", true),
+                   peg$c5 = "digraph",
+                   peg$c6 = peg$literalExpectation("digraph", true),
+                   peg$c7 = "{",
+                   peg$c8 = peg$literalExpectation("{", false),
+                   peg$c9 = "}",
+                   peg$c10 = peg$literalExpectation("}", false),
+                   peg$c11 = function(strict, type, id, children) {
+                       var ret = {type:type.toLowerCase(), children:children};
+                       if (strict) { ret.strict = true }
+                       if (id) { ret.id = id }
+                       return ret;
+                   },
+                   peg$c12 = ";",
+                   peg$c13 = peg$literalExpectation(";", false),
+                   peg$c14 = function(s, v) {return [s].concat(v||[]);},
+                   peg$c15 = function(v) { return v; },
+                   peg$c16 = "=",
+                   peg$c17 = peg$literalExpectation("=", false),
+                   peg$c18 = function(left, right) { return {type:'attr_stmt', target:'graph', attr_list:[{type:'attr', id:left, eq:right}]}; },
+                   peg$c19 = "node",
+                   peg$c20 = peg$literalExpectation("node", true),
+                   peg$c21 = "edge",
+                   peg$c22 = peg$literalExpectation("edge", true),
+                   peg$c23 = function(target, attr) { return {type:'attr_stmt', target:target, attr_list:attr}; },
+                   peg$c24 = "[",
+                   peg$c25 = peg$literalExpectation("[", false),
+                   peg$c26 = "]",
+                   peg$c27 = peg$literalExpectation("]", false),
+                   peg$c28 = function(list, rest) { return (list || []).concat(rest || []); },
+                   peg$c29 = function(elementType, elementValue) {return elementValue},
+                   peg$c30 = ",",
+                   peg$c31 = peg$literalExpectation(",", false),
+                   peg$c32 = function(elementType, eq, rest) {
+                       var attributes = [];
+   
+                       function createAttribute(name, value) {
+                           return {
+                               type: 'attr',
+                               name: name,
+                               value: value
+                           };
+                       }
+   
+                       function trim(str){
+                           return String(str).replace(/^\s+|\s+$/g, '');
+                       }
+   
+                       if (elementType !== 'label') {
+                           attributes.push(createAttribute(elementType, eq || null));
+                       } else {
+                           var eq = eq.replace(/\\n/ig, "\n").split(/\n/);
+                           var newAttributes = eq.map(function(pair) {
+                               var values = pair.split(':');
+                               var newName = values[0] ? trim(values[0]) : values[0];
+                               var newValue = values[1] ? trim(values[1]) : values[1];
+                               if (newName === 'dims'){
+                                   newValue = JSON.parse(newValue);
+                               }
+                               return createAttribute(newName, newValue);
+                           });
+                           newAttributes[0].value = newAttributes[0].name;
+                           newAttributes[0].name = 'name';
+   
+                           Array.prototype.push.apply(attributes, newAttributes);
+                       }
+   
+                       return attributes.concat(rest || []);
+                   },
+                   peg$c33 = function(id, rhs, attr) {
+                       var edge_list = [id];
+                       edge_list = edge_list.concat(rhs.map(function(v){return v.id}));
+                       return {type:'edge_stmt', edge_list:edge_list, attr_list:attr||[]};
+                   },
+                   peg$c34 = "->",
+                   peg$c35 = peg$literalExpectation("->", false),
+                   peg$c36 = "--",
+                   peg$c37 = peg$literalExpectation("--", false),
+                   peg$c38 = function(edgeop, id, rest) {return [{type:'edgeRHS', edgeop:edgeop, id:id}].concat(rest || []); },
+                   peg$c39 = function(id, attr) { return {type:'node_stmt', node_id:id, attr_list:attr || []}; },
+                   peg$c40 = function(id, port) { return port ? {type:'node_id', id:id, port:port} : {type:'node_id', id:id}; },
+                   peg$c41 = peg$otherExpectation("port"),
+                   peg$c42 = ":",
+                   peg$c43 = peg$literalExpectation(":", false),
+                   peg$c44 = function(id, pt) {return pt},
+                   peg$c45 = function(id, pt) { return {type:'port', id:id, compass_pt:pt || null}; },
+                   peg$c46 = function(pt) {return {type:'port', compass_pt:pt||null}},
+                   peg$c47 = "subgraph",
+                   peg$c48 = peg$literalExpectation("subgraph", true),
+                   peg$c49 = function(id) {return id ? {type:'subgraph', id:id} : {type:'subgraph'}},
+                   peg$c50 = function(g, s) { g=g||{type:'subgraph'}; g.children = s||[]; return g; },
+                   peg$c51 = function(id) { return {type:'subgraph', id:id, children:[]}; },
+                   peg$c52 = "n",
+                   peg$c53 = peg$literalExpectation("n", false),
+                   peg$c54 = "ne",
+                   peg$c55 = peg$literalExpectation("ne", false),
+                   peg$c56 = "e",
+                   peg$c57 = peg$literalExpectation("e", false),
+                   peg$c58 = "se",
+                   peg$c59 = peg$literalExpectation("se", false),
+                   peg$c60 = "s",
+                   peg$c61 = peg$literalExpectation("s", false),
+                   peg$c62 = "sw",
+                   peg$c63 = peg$literalExpectation("sw", false),
+                   peg$c64 = "w",
+                   peg$c65 = peg$literalExpectation("w", false),
+                   peg$c66 = "nw",
+                   peg$c67 = peg$literalExpectation("nw", false),
+                   peg$c68 = /^[a-zA-Z_]/,
+                   peg$c69 = peg$classExpectation([["a", "z"], ["A", "Z"], "_"], false, false),
+                   peg$c70 = /^[a-zA-Z0-9_]/,
+                   peg$c71 = peg$classExpectation([["a", "z"], ["A", "Z"], ["0", "9"], "_"], false, false),
+                   peg$c72 = function(v) { return v[0]+v[1].join(''); },
+                   peg$c73 = peg$otherExpectation("NUMBER"),
+                   peg$c74 = "-",
+                   peg$c75 = peg$literalExpectation("-", false),
+                   peg$c76 = ".",
+                   peg$c77 = peg$literalExpectation(".", false),
+                   peg$c78 = /^[0-9]/,
+                   peg$c79 = peg$classExpectation([["0", "9"]], false, false),
+                   peg$c80 = function(n) {
+                       return parseFloat(text());
+                   },
+                   peg$c81 = "<",
+                   peg$c82 = peg$literalExpectation("<", false),
+                   peg$c83 = ">",
+                   peg$c84 = peg$literalExpectation(">", false),
+                   peg$c85 = function(v) { return '<' + v.join('') + '>'; },
+                   peg$c86 = peg$anyExpectation(),
+                   peg$c87 = function(v) {return v},
+                   peg$c88 = function(v) { return v.join(""); },
+                   peg$c89 = "\"",
+                   peg$c90 = peg$literalExpectation("\"", false),
+                   peg$c91 = function() {return "";},
+                   peg$c92 = "\\",
+                   peg$c93 = peg$literalExpectation("\\", false),
+                   peg$c94 = "+",
+                   peg$c95 = peg$literalExpectation("+", false),
+                   peg$c96 = function(v2, v) {return v},
+                   peg$c97 = function(v2, rest) { return rest === null ? v2[1] : (v2[1] + rest); },
+                   peg$c98 = function(chars) { return chars.join(""); },
+                   peg$c99 = /^[^"\\\0-\x1F\x7F]/,
+                   peg$c100 = peg$classExpectation(["\"", "\\", ["\0", "\x1F"], "\x7F"], true, false),
+                   peg$c101 = "\\\"",
+                   peg$c102 = peg$literalExpectation("\\\"", false),
+                   peg$c103 = function() { return '"'; },
+                   peg$c104 = function() { return ""; },
+                   peg$c105 = function() { return '\\'; },
+                   peg$c106 = peg$otherExpectation("COMMENT"),
+                   peg$c107 = peg$otherExpectation("BLOCK_COMMENT"),
+                   peg$c108 = "/*",
+                   peg$c109 = peg$literalExpectation("/*", false),
+                   peg$c110 = "*/",
+                   peg$c111 = peg$literalExpectation("*/", false),
+                   peg$c112 = function(v) {return v;},
+                   peg$c113 = function(v) { return v.join('') },
+                   peg$c114 = peg$otherExpectation("C_COMMENT"),
+                   peg$c115 = "//",
+                   peg$c116 = peg$literalExpectation("//", false),
+                   peg$c117 = /^[\n]/,
+                   peg$c118 = peg$classExpectation(["\n"], false, false),
+                   peg$c119 = function(v) { return v.join(''); },
+                   peg$c120 = peg$otherExpectation("MACRO_COMMENT"),
+                   peg$c121 = "#",
+                   peg$c122 = peg$literalExpectation("#", false),
+                   peg$c123 = peg$otherExpectation("WHITESPACE"),
+                   peg$c124 = /^[\n\r]/,
+                   peg$c125 = peg$classExpectation(["\n", "\r"], false, false),
+                   peg$c126 = /^[ \t]/,
+                   peg$c127 = peg$classExpectation([" ", "\t"], false, false),
+   
+                   peg$currPos          = 0,
+                   peg$savedPos         = 0,
+                   peg$posDetailsCache  = [{ line: 1, column: 1 }],
+                   peg$maxFailPos       = 0,
+                   peg$maxFailExpected  = [],
+                   peg$silentFails      = 0,
+   
+                   peg$result;
+   
+               if ("startRule" in options) {
+                   if (!(options.startRule in peg$startRuleFunctions)) {
+                       throw new Error("Can't start parsing from rule \"" + options.startRule + "\".");
+                   }
+   
+                   peg$startRuleFunction = peg$startRuleFunctions[options.startRule];
+               }
+   
+               function text() {
+                   return input.substring(peg$savedPos, peg$currPos);
+               }
+   
+               function location() {
+                   return peg$computeLocation(peg$savedPos, peg$currPos);
+               }
+   
+               function expected(description, location) {
+                   location = location !== void 0 ? location : peg$computeLocation(peg$savedPos, peg$currPos)
+   
+                   throw peg$buildStructuredError(
+                       [peg$otherExpectation(description)],
+                       input.substring(peg$savedPos, peg$currPos),
+                       location
+                   );
+               }
+   
+               function error(message, location) {
+                   location = location !== void 0 ? location : peg$computeLocation(peg$savedPos, peg$currPos)
+   
+                   throw peg$buildSimpleError(message, location);
+               }
+   
+               function peg$literalExpectation(text, ignoreCase) {
+                   return { type: "literal", text: text, ignoreCase: ignoreCase };
+               }
+   
+               function peg$classExpectation(parts, inverted, ignoreCase) {
+                   return { type: "class", parts: parts, inverted: inverted, ignoreCase: ignoreCase };
+               }
+   
+               function peg$anyExpectation() {
+                   return { type: "any" };
+               }
+   
+               function peg$endExpectation() {
+                   return { type: "end" };
+               }
+   
+               function peg$otherExpectation(description) {
+                   return { type: "other", description: description };
+               }
+   
+               function peg$computePosDetails(pos) {
+                   var details = peg$posDetailsCache[pos], p;
+   
+                   if (details) {
+                       return details;
+                   } else {
+                       p = pos - 1;
+                       while (!peg$posDetailsCache[p]) {
+                           p--;
+                       }
+   
+                       details = peg$posDetailsCache[p];
+                       details = {
+                           line:   details.line,
+                           column: details.column
+                       };
+   
+                       while (p < pos) {
+                           if (input.charCodeAt(p) === 10) {
+                               details.line++;
+                               details.column = 1;
+                           } else {
+                               details.column++;
+                           }
+   
+                           p++;
+                       }
+   
+                       peg$posDetailsCache[pos] = details;
+                       return details;
+                   }
+               }
+   
+               function peg$computeLocation(startPos, endPos) {
+                   var startPosDetails = peg$computePosDetails(startPos),
+                       endPosDetails   = peg$computePosDetails(endPos);
+   
+                   return {
+                       start: {
+                           offset: startPos,
+                           line:   startPosDetails.line,
+                           column: startPosDetails.column
+                       },
+                       end: {
+                           offset: endPos,
+                           line:   endPosDetails.line,
+                           column: endPosDetails.column
+                       }
+                   };
+               }
+   
+               function peg$fail(expected) {
+                   if (peg$currPos < peg$maxFailPos) { return; }
+   
+                   if (peg$currPos > peg$maxFailPos) {
+                       peg$maxFailPos = peg$currPos;
+                       peg$maxFailExpected = [];
+                   }
+   
+                   peg$maxFailExpected.push(expected);
+               }
+   
+               function peg$buildSimpleError(message, location) {
+                   return new peg$SyntaxError(message, null, null, location);
+               }
+   
+               function peg$buildStructuredError(expected, found, location) {
+                   return new peg$SyntaxError(
+                       peg$SyntaxError.buildMessage(expected, found),
+                       expected,
+                       found,
+                       location
+                   );
+               }
+   
+               function peg$parsestart() {
+                   var s0;
+   
+                   s0 = peg$parsegraph();
+   
+                   return s0;
+               }
+   
+               function peg$parsegraph() {
+                   var s0, s1, s2, s3, s4, s5, s6, s7, s8, s9, s10, s11;
+   
+                   peg$silentFails++;
+                   s0 = peg$currPos;
+                   s1 = peg$parse_();
+                   if (s1 !== peg$FAILED) {
+                       if (input.substr(peg$currPos, 6).toLowerCase() === peg$c1) {
+                           s2 = input.substr(peg$currPos, 6);
+                           peg$currPos += 6;
+                       } else {
+                           s2 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c2); }
+                       }
+                       if (s2 === peg$FAILED) {
+                           s2 = null;
+                       }
+                       if (s2 !== peg$FAILED) {
+                           s3 = peg$parse_();
+                           if (s3 !== peg$FAILED) {
+                               if (input.substr(peg$currPos, 5).toLowerCase() === peg$c3) {
+                                   s4 = input.substr(peg$currPos, 5);
+                                   peg$currPos += 5;
+                               } else {
+                                   s4 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c4); }
+                               }
+                               if (s4 === peg$FAILED) {
+                                   if (input.substr(peg$currPos, 7).toLowerCase() === peg$c5) {
+                                       s4 = input.substr(peg$currPos, 7);
+                                       peg$currPos += 7;
+                                   } else {
+                                       s4 = peg$FAILED;
+                                       if (peg$silentFails === 0) { peg$fail(peg$c6); }
+                                   }
+                               }
+                               if (s4 !== peg$FAILED) {
+                                   s5 = peg$parse_();
+                                   if (s5 !== peg$FAILED) {
+                                       s6 = peg$parseID();
+                                       if (s6 === peg$FAILED) {
+                                           s6 = null;
+                                       }
+                                       if (s6 !== peg$FAILED) {
+                                           s7 = peg$parse_();
+                                           if (s7 !== peg$FAILED) {
+                                               if (input.charCodeAt(peg$currPos) === 123) {
+                                                   s8 = peg$c7;
+                                                   peg$currPos++;
+                                               } else {
+                                                   s8 = peg$FAILED;
+                                                   if (peg$silentFails === 0) { peg$fail(peg$c8); }
+                                               }
+                                               if (s8 !== peg$FAILED) {
+                                                   s9 = peg$parsestmt_list();
+                                                   if (s9 !== peg$FAILED) {
+                                                       if (input.charCodeAt(peg$currPos) === 125) {
+                                                           s10 = peg$c9;
+                                                           peg$currPos++;
+                                                       } else {
+                                                           s10 = peg$FAILED;
+                                                           if (peg$silentFails === 0) { peg$fail(peg$c10); }
+                                                       }
+                                                       if (s10 !== peg$FAILED) {
+                                                           s11 = peg$parse_();
+                                                           if (s11 !== peg$FAILED) {
+                                                               peg$savedPos = s0;
+                                                               s1 = peg$c11(s2, s4, s6, s9);
+                                                               s0 = s1;
+                                                           } else {
+                                                               peg$currPos = s0;
+                                                               s0 = peg$FAILED;
+                                                           }
+                                                       } else {
+                                                           peg$currPos = s0;
+                                                           s0 = peg$FAILED;
+                                                       }
+                                                   } else {
+                                                       peg$currPos = s0;
+                                                       s0 = peg$FAILED;
+                                                   }
+                                               } else {
+                                                   peg$currPos = s0;
+                                                   s0 = peg$FAILED;
+                                               }
+                                           } else {
+                                               peg$currPos = s0;
+                                               s0 = peg$FAILED;
+                                           }
+                                       } else {
+                                           peg$currPos = s0;
+                                           s0 = peg$FAILED;
+                                       }
+                                   } else {
+                                       peg$currPos = s0;
+                                       s0 = peg$FAILED;
+                                   }
+                               } else {
+                                   peg$currPos = s0;
+                                   s0 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+                   peg$silentFails--;
+                   if (s0 === peg$FAILED) {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c0); }
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parsestmt_list() {
+                   var s0, s1, s2, s3, s4, s5, s6, s7;
+   
+                   s0 = peg$currPos;
+                   s1 = peg$parse_();
+                   if (s1 !== peg$FAILED) {
+                       s2 = peg$currPos;
+                       s3 = peg$parsestmt();
+                       if (s3 !== peg$FAILED) {
+                           s4 = peg$parse_();
+                           if (s4 !== peg$FAILED) {
+                               if (input.charCodeAt(peg$currPos) === 59) {
+                                   s5 = peg$c12;
+                                   peg$currPos++;
+                               } else {
+                                   s5 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c13); }
+                               }
+                               if (s5 === peg$FAILED) {
+                                   s5 = null;
+                               }
+                               if (s5 !== peg$FAILED) {
+                                   s6 = peg$parse_();
+                                   if (s6 !== peg$FAILED) {
+                                       s7 = peg$parsestmt_list();
+                                       if (s7 === peg$FAILED) {
+                                           s7 = null;
+                                       }
+                                       if (s7 !== peg$FAILED) {
+                                           peg$savedPos = s2;
+                                           s3 = peg$c14(s3, s7);
+                                           s2 = s3;
+                                       } else {
+                                           peg$currPos = s2;
+                                           s2 = peg$FAILED;
+                                       }
+                                   } else {
+                                       peg$currPos = s2;
+                                       s2 = peg$FAILED;
+                                   }
+                               } else {
+                                   peg$currPos = s2;
+                                   s2 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s2;
+                               s2 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s2;
+                           s2 = peg$FAILED;
+                       }
+                       if (s2 === peg$FAILED) {
+                           s2 = null;
+                       }
+                       if (s2 !== peg$FAILED) {
+                           peg$savedPos = s0;
+                           s1 = peg$c15(s2);
+                           s0 = s1;
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parsestmt() {
+                   var s0, s1, s2, s3, s4, s5;
+   
+                   s0 = peg$currPos;
+                   s1 = peg$parseID();
+                   if (s1 !== peg$FAILED) {
+                       s2 = peg$parse_();
+                       if (s2 !== peg$FAILED) {
+                           if (input.charCodeAt(peg$currPos) === 61) {
+                               s3 = peg$c16;
+                               peg$currPos++;
+                           } else {
+                               s3 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c17); }
+                           }
+                           if (s3 !== peg$FAILED) {
+                               s4 = peg$parse_();
+                               if (s4 !== peg$FAILED) {
+                                   s5 = peg$parseID();
+                                   if (s5 !== peg$FAILED) {
+                                       peg$savedPos = s0;
+                                       s1 = peg$c18(s1, s5);
+                                       s0 = s1;
+                                   } else {
+                                       peg$currPos = s0;
+                                       s0 = peg$FAILED;
+                                   }
+                               } else {
+                                   peg$currPos = s0;
+                                   s0 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+                   if (s0 === peg$FAILED) {
+                       s0 = peg$parseattr_stmt();
+                       if (s0 === peg$FAILED) {
+                           s0 = peg$parseedge_stmt();
+                           if (s0 === peg$FAILED) {
+                               s0 = peg$parsesubgraph();
+                               if (s0 === peg$FAILED) {
+                                   s0 = peg$parsenode_stmt();
+                                   if (s0 === peg$FAILED) {
+                                       s0 = peg$currPos;
+                                       s1 = peg$parseID();
+                                       if (s1 !== peg$FAILED) {
+                                           if (input.charCodeAt(peg$currPos) === 61) {
+                                               s2 = peg$c16;
+                                               peg$currPos++;
+                                           } else {
+                                               s2 = peg$FAILED;
+                                               if (peg$silentFails === 0) { peg$fail(peg$c17); }
+                                           }
+                                           if (s2 !== peg$FAILED) {
+                                               s3 = peg$parseID();
+                                               if (s3 !== peg$FAILED) {
+                                                   s1 = [s1, s2, s3];
+                                                   s0 = s1;
+                                               } else {
+                                                   peg$currPos = s0;
+                                                   s0 = peg$FAILED;
+                                               }
+                                           } else {
+                                               peg$currPos = s0;
+                                               s0 = peg$FAILED;
+                                           }
+                                       } else {
+                                           peg$currPos = s0;
+                                           s0 = peg$FAILED;
+                                       }
+                                   }
+                               }
+                           }
+                       }
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parseattr_stmt() {
+                   var s0, s1, s2;
+   
+                   s0 = peg$currPos;
+                   if (input.substr(peg$currPos, 5).toLowerCase() === peg$c3) {
+                       s1 = input.substr(peg$currPos, 5);
+                       peg$currPos += 5;
+                   } else {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c4); }
+                   }
+                   if (s1 === peg$FAILED) {
+                       if (input.substr(peg$currPos, 4).toLowerCase() === peg$c19) {
+                           s1 = input.substr(peg$currPos, 4);
+                           peg$currPos += 4;
+                       } else {
+                           s1 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c20); }
+                       }
+                       if (s1 === peg$FAILED) {
+                           if (input.substr(peg$currPos, 4).toLowerCase() === peg$c21) {
+                               s1 = input.substr(peg$currPos, 4);
+                               peg$currPos += 4;
+                           } else {
+                               s1 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c22); }
+                           }
+                       }
+                   }
+                   if (s1 !== peg$FAILED) {
+                       s2 = peg$parseattr_list();
+                       if (s2 !== peg$FAILED) {
+                           peg$savedPos = s0;
+                           s1 = peg$c23(s1, s2);
+                           s0 = s1;
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parseattr_list() {
+                   var s0, s1, s2, s3, s4, s5, s6;
+   
+                   s0 = peg$currPos;
+                   s1 = peg$parse_();
+                   if (s1 !== peg$FAILED) {
+                       if (input.charCodeAt(peg$currPos) === 91) {
+                           s2 = peg$c24;
+                           peg$currPos++;
+                       } else {
+                           s2 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c25); }
+                       }
+                       if (s2 !== peg$FAILED) {
+                           s3 = peg$parsea_list();
+                           if (s3 === peg$FAILED) {
+                               s3 = null;
+                           }
+                           if (s3 !== peg$FAILED) {
+                               if (input.charCodeAt(peg$currPos) === 93) {
+                                   s4 = peg$c26;
+                                   peg$currPos++;
+                               } else {
+                                   s4 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c27); }
+                               }
+                               if (s4 !== peg$FAILED) {
+                                   s5 = peg$parse_();
+                                   if (s5 !== peg$FAILED) {
+                                       s6 = peg$parseattr_list();
+                                       if (s6 === peg$FAILED) {
+                                           s6 = null;
+                                       }
+                                       if (s6 !== peg$FAILED) {
+                                           peg$savedPos = s0;
+                                           s1 = peg$c28(s3, s6);
+                                           s0 = s1;
+                                       } else {
+                                           peg$currPos = s0;
+                                           s0 = peg$FAILED;
+                                       }
+                                   } else {
+                                       peg$currPos = s0;
+                                       s0 = peg$FAILED;
+                                   }
+                               } else {
+                                   peg$currPos = s0;
+                                   s0 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parsea_list() {
+                   var s0, s1, s2, s3, s4, s5, s6, s7;
+   
+                   s0 = peg$currPos;
+                   s1 = peg$parse_();
+                   if (s1 !== peg$FAILED) {
+                       s2 = peg$parseID();
+                       if (s2 !== peg$FAILED) {
+                           s3 = peg$currPos;
+                           s4 = peg$parse_();
+                           if (s4 !== peg$FAILED) {
+                               if (input.charCodeAt(peg$currPos) === 61) {
+                                   s5 = peg$c16;
+                                   peg$currPos++;
+                               } else {
+                                   s5 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c17); }
+                               }
+                               if (s5 !== peg$FAILED) {
+                                   s6 = peg$parse_();
+                                   if (s6 !== peg$FAILED) {
+                                       s7 = peg$parseID();
+                                       if (s7 !== peg$FAILED) {
+                                           peg$savedPos = s3;
+                                           s4 = peg$c29(s2, s7);
+                                           s3 = s4;
+                                       } else {
+                                           peg$currPos = s3;
+                                           s3 = peg$FAILED;
+                                       }
+                                   } else {
+                                       peg$currPos = s3;
+                                       s3 = peg$FAILED;
+                                   }
+                               } else {
+                                   peg$currPos = s3;
+                                   s3 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s3;
+                               s3 = peg$FAILED;
+                           }
+                           if (s3 === peg$FAILED) {
+                               s3 = null;
+                           }
+                           if (s3 !== peg$FAILED) {
+                               s4 = peg$parse_();
+                               if (s4 !== peg$FAILED) {
+                                   if (input.charCodeAt(peg$currPos) === 44) {
+                                       s5 = peg$c30;
+                                       peg$currPos++;
+                                   } else {
+                                       s5 = peg$FAILED;
+                                       if (peg$silentFails === 0) { peg$fail(peg$c31); }
+                                   }
+                                   if (s5 === peg$FAILED) {
+                                       s5 = null;
+                                   }
+                                   if (s5 !== peg$FAILED) {
+                                       s6 = peg$parsea_list();
+                                       if (s6 === peg$FAILED) {
+                                           s6 = null;
+                                       }
+                                       if (s6 !== peg$FAILED) {
+                                           peg$savedPos = s0;
+                                           s1 = peg$c32(s2, s3, s6);
+                                           s0 = s1;
+                                       } else {
+                                           peg$currPos = s0;
+                                           s0 = peg$FAILED;
+                                       }
+                                   } else {
+                                       peg$currPos = s0;
+                                       s0 = peg$FAILED;
+                                   }
+                               } else {
+                                   peg$currPos = s0;
+                                   s0 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parseedge_stmt() {
+                   var s0, s1, s2, s3;
+   
+                   s0 = peg$currPos;
+                   s1 = peg$parsesubgraph();
+                   if (s1 === peg$FAILED) {
+                       s1 = peg$parsenode_id();
+                   }
+                   if (s1 !== peg$FAILED) {
+                       s2 = peg$parseedgeRHS();
+                       if (s2 !== peg$FAILED) {
+                           s3 = peg$parseattr_list();
+                           if (s3 === peg$FAILED) {
+                               s3 = null;
+                           }
+                           if (s3 !== peg$FAILED) {
+                               peg$savedPos = s0;
+                               s1 = peg$c33(s1, s2, s3);
+                               s0 = s1;
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parseedgeRHS() {
+                   var s0, s1, s2, s3, s4, s5, s6;
+   
+                   s0 = peg$currPos;
+                   s1 = peg$parse_();
+                   if (s1 !== peg$FAILED) {
+                       if (input.substr(peg$currPos, 2) === peg$c34) {
+                           s2 = peg$c34;
+                           peg$currPos += 2;
+                       } else {
+                           s2 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c35); }
+                       }
+                       if (s2 === peg$FAILED) {
+                           if (input.substr(peg$currPos, 2) === peg$c36) {
+                               s2 = peg$c36;
+                               peg$currPos += 2;
+                           } else {
+                               s2 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c37); }
+                           }
+                       }
+                       if (s2 !== peg$FAILED) {
+                           s3 = peg$parse_();
+                           if (s3 !== peg$FAILED) {
+                               s4 = peg$parsesubgraph();
+                               if (s4 === peg$FAILED) {
+                                   s4 = peg$parsenode_id();
+                               }
+                               if (s4 !== peg$FAILED) {
+                                   s5 = peg$parse_();
+                                   if (s5 !== peg$FAILED) {
+                                       s6 = peg$parseedgeRHS();
+                                       if (s6 === peg$FAILED) {
+                                           s6 = null;
+                                       }
+                                       if (s6 !== peg$FAILED) {
+                                           peg$savedPos = s0;
+                                           s1 = peg$c38(s2, s4, s6);
+                                           s0 = s1;
+                                       } else {
+                                           peg$currPos = s0;
+                                           s0 = peg$FAILED;
+                                       }
+                                   } else {
+                                       peg$currPos = s0;
+                                       s0 = peg$FAILED;
+                                   }
+                               } else {
+                                   peg$currPos = s0;
+                                   s0 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parsenode_stmt() {
+                   var s0, s1, s2;
+   
+                   s0 = peg$currPos;
+                   s1 = peg$parsenode_id();
+                   if (s1 !== peg$FAILED) {
+                       s2 = peg$parseattr_list();
+                       if (s2 === peg$FAILED) {
+                           s2 = null;
+                       }
+                       if (s2 !== peg$FAILED) {
+                           peg$savedPos = s0;
+                           s1 = peg$c39(s1, s2);
+                           s0 = s1;
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parsenode_id() {
+                   var s0, s1, s2;
+   
+                   s0 = peg$currPos;
+                   s1 = peg$parseID();
+                   if (s1 !== peg$FAILED) {
+                       s2 = peg$parseport();
+                       if (s2 === peg$FAILED) {
+                           s2 = null;
+                       }
+                       if (s2 !== peg$FAILED) {
+                           peg$savedPos = s0;
+                           s1 = peg$c40(s1, s2);
+                           s0 = s1;
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parseport() {
+                   var s0, s1, s2, s3, s4, s5;
+   
+                   peg$silentFails++;
+                   s0 = peg$currPos;
+                   if (input.charCodeAt(peg$currPos) === 58) {
+                       s1 = peg$c42;
+                       peg$currPos++;
+                   } else {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c43); }
+                   }
+                   if (s1 !== peg$FAILED) {
+                       s2 = peg$parseID();
+                       if (s2 !== peg$FAILED) {
+                           s3 = peg$currPos;
+                           if (input.charCodeAt(peg$currPos) === 58) {
+                               s4 = peg$c42;
+                               peg$currPos++;
+                           } else {
+                               s4 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c43); }
+                           }
+                           if (s4 !== peg$FAILED) {
+                               s5 = peg$parsecompass_pt();
+                               if (s5 !== peg$FAILED) {
+                                   peg$savedPos = s3;
+                                   s4 = peg$c44(s2, s5);
+                                   s3 = s4;
+                               } else {
+                                   peg$currPos = s3;
+                                   s3 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s3;
+                               s3 = peg$FAILED;
+                           }
+                           if (s3 === peg$FAILED) {
+                               s3 = null;
+                           }
+                           if (s3 !== peg$FAILED) {
+                               peg$savedPos = s0;
+                               s1 = peg$c45(s2, s3);
+                               s0 = s1;
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+                   if (s0 === peg$FAILED) {
+                       s0 = peg$currPos;
+                       if (input.charCodeAt(peg$currPos) === 58) {
+                           s1 = peg$c42;
+                           peg$currPos++;
+                       } else {
+                           s1 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c43); }
+                       }
+                       if (s1 !== peg$FAILED) {
+                           s2 = peg$parsecompass_pt();
+                           if (s2 !== peg$FAILED) {
+                               peg$savedPos = s0;
+                               s1 = peg$c46(s2);
+                               s0 = s1;
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   }
+                   peg$silentFails--;
+                   if (s0 === peg$FAILED) {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c41); }
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parsesubgraph() {
+                   var s0, s1, s2, s3, s4, s5;
+   
+                   s0 = peg$currPos;
+                   s1 = peg$currPos;
+                   if (input.substr(peg$currPos, 8).toLowerCase() === peg$c47) {
+                       s2 = input.substr(peg$currPos, 8);
+                       peg$currPos += 8;
+                   } else {
+                       s2 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c48); }
+                   }
+                   if (s2 !== peg$FAILED) {
+                       s3 = peg$parse_();
+                       if (s3 !== peg$FAILED) {
+                           s4 = peg$parseID();
+                           if (s4 === peg$FAILED) {
+                               s4 = null;
+                           }
+                           if (s4 !== peg$FAILED) {
+                               s5 = peg$parse_();
+                               if (s5 !== peg$FAILED) {
+                                   peg$savedPos = s1;
+                                   s2 = peg$c49(s4);
+                                   s1 = s2;
+                               } else {
+                                   peg$currPos = s1;
+                                   s1 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s1;
+                               s1 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s1;
+                           s1 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s1;
+                       s1 = peg$FAILED;
+                   }
+                   if (s1 === peg$FAILED) {
+                       s1 = null;
+                   }
+                   if (s1 !== peg$FAILED) {
+                       if (input.charCodeAt(peg$currPos) === 123) {
+                           s2 = peg$c7;
+                           peg$currPos++;
+                       } else {
+                           s2 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c8); }
+                       }
+                       if (s2 !== peg$FAILED) {
+                           s3 = peg$parsestmt_list();
+                           if (s3 !== peg$FAILED) {
+                               if (input.charCodeAt(peg$currPos) === 125) {
+                                   s4 = peg$c9;
+                                   peg$currPos++;
+                               } else {
+                                   s4 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c10); }
+                               }
+                               if (s4 !== peg$FAILED) {
+                                   peg$savedPos = s0;
+                                   s1 = peg$c50(s1, s3);
+                                   s0 = s1;
+                               } else {
+                                   peg$currPos = s0;
+                                   s0 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+                   if (s0 === peg$FAILED) {
+                       s0 = peg$currPos;
+                       if (input.substr(peg$currPos, 8).toLowerCase() === peg$c47) {
+                           s1 = input.substr(peg$currPos, 8);
+                           peg$currPos += 8;
+                       } else {
+                           s1 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c48); }
+                       }
+                       if (s1 !== peg$FAILED) {
+                           s2 = peg$parse_();
+                           if (s2 !== peg$FAILED) {
+                               s3 = peg$parseID();
+                               if (s3 !== peg$FAILED) {
+                                   peg$savedPos = s0;
+                                   s1 = peg$c51(s3);
+                                   s0 = s1;
+                               } else {
+                                   peg$currPos = s0;
+                                   s0 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parsecompass_pt() {
+                   var s0;
+   
+                   if (input.charCodeAt(peg$currPos) === 110) {
+                       s0 = peg$c52;
+                       peg$currPos++;
+                   } else {
+                       s0 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c53); }
+                   }
+                   if (s0 === peg$FAILED) {
+                       if (input.substr(peg$currPos, 2) === peg$c54) {
+                           s0 = peg$c54;
+                           peg$currPos += 2;
+                       } else {
+                           s0 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c55); }
+                       }
+                       if (s0 === peg$FAILED) {
+                           if (input.charCodeAt(peg$currPos) === 101) {
+                               s0 = peg$c56;
+                               peg$currPos++;
+                           } else {
+                               s0 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c57); }
+                           }
+                           if (s0 === peg$FAILED) {
+                               if (input.substr(peg$currPos, 2) === peg$c58) {
+                                   s0 = peg$c58;
+                                   peg$currPos += 2;
+                               } else {
+                                   s0 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c59); }
+                               }
+                               if (s0 === peg$FAILED) {
+                                   if (input.charCodeAt(peg$currPos) === 115) {
+                                       s0 = peg$c60;
+                                       peg$currPos++;
+                                   } else {
+                                       s0 = peg$FAILED;
+                                       if (peg$silentFails === 0) { peg$fail(peg$c61); }
+                                   }
+                                   if (s0 === peg$FAILED) {
+                                       if (input.substr(peg$currPos, 2) === peg$c62) {
+                                           s0 = peg$c62;
+                                           peg$currPos += 2;
+                                       } else {
+                                           s0 = peg$FAILED;
+                                           if (peg$silentFails === 0) { peg$fail(peg$c63); }
+                                       }
+                                       if (s0 === peg$FAILED) {
+                                           if (input.charCodeAt(peg$currPos) === 119) {
+                                               s0 = peg$c64;
+                                               peg$currPos++;
+                                           } else {
+                                               s0 = peg$FAILED;
+                                               if (peg$silentFails === 0) { peg$fail(peg$c65); }
+                                           }
+                                           if (s0 === peg$FAILED) {
+                                               if (input.substr(peg$currPos, 2) === peg$c66) {
+                                                   s0 = peg$c66;
+                                                   peg$currPos += 2;
+                                               } else {
+                                                   s0 = peg$FAILED;
+                                                   if (peg$silentFails === 0) { peg$fail(peg$c67); }
+                                               }
+                                           }
+                                       }
+                                   }
+                               }
+                           }
+                       }
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parseID() {
+                   var s0;
+   
+                   s0 = peg$parseSTRING();
+                   if (s0 === peg$FAILED) {
+                       s0 = peg$parseNUMBER();
+                       if (s0 === peg$FAILED) {
+                           s0 = peg$parseQUOTED_STRING();
+                       }
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parseSTRING() {
+                   var s0, s1, s2, s3, s4;
+   
+                   s0 = peg$currPos;
+                   s1 = peg$currPos;
+                   if (peg$c68.test(input.charAt(peg$currPos))) {
+                       s2 = input.charAt(peg$currPos);
+                       peg$currPos++;
+                   } else {
+                       s2 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c69); }
+                   }
+                   if (s2 !== peg$FAILED) {
+                       s3 = [];
+                       if (peg$c70.test(input.charAt(peg$currPos))) {
+                           s4 = input.charAt(peg$currPos);
+                           peg$currPos++;
+                       } else {
+                           s4 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c71); }
+                       }
+                       while (s4 !== peg$FAILED) {
+                           s3.push(s4);
+                           if (peg$c70.test(input.charAt(peg$currPos))) {
+                               s4 = input.charAt(peg$currPos);
+                               peg$currPos++;
+                           } else {
+                               s4 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c71); }
+                           }
+                       }
+                       if (s3 !== peg$FAILED) {
+                           s2 = [s2, s3];
+                           s1 = s2;
+                       } else {
+                           peg$currPos = s1;
+                           s1 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s1;
+                       s1 = peg$FAILED;
+                   }
+                   if (s1 !== peg$FAILED) {
+                       peg$savedPos = s0;
+                       s1 = peg$c72(s1);
+                   }
+                   s0 = s1;
+   
+                   return s0;
+               }
+   
+               function peg$parseNUMBER() {
+                   var s0, s1, s2, s3, s4, s5, s6, s7, s8;
+   
+                   peg$silentFails++;
+                   s0 = peg$currPos;
+                   s1 = peg$currPos;
+                   if (input.charCodeAt(peg$currPos) === 45) {
+                       s2 = peg$c74;
+                       peg$currPos++;
+                   } else {
+                       s2 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c75); }
+                   }
+                   if (s2 === peg$FAILED) {
+                       s2 = null;
+                   }
+                   if (s2 !== peg$FAILED) {
+                       s3 = peg$currPos;
+                       if (input.charCodeAt(peg$currPos) === 46) {
+                           s4 = peg$c76;
+                           peg$currPos++;
+                       } else {
+                           s4 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c77); }
+                       }
+                       if (s4 !== peg$FAILED) {
+                           s5 = [];
+                           if (peg$c78.test(input.charAt(peg$currPos))) {
+                               s6 = input.charAt(peg$currPos);
+                               peg$currPos++;
+                           } else {
+                               s6 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c79); }
+                           }
+                           if (s6 !== peg$FAILED) {
+                               while (s6 !== peg$FAILED) {
+                                   s5.push(s6);
+                                   if (peg$c78.test(input.charAt(peg$currPos))) {
+                                       s6 = input.charAt(peg$currPos);
+                                       peg$currPos++;
+                                   } else {
+                                       s6 = peg$FAILED;
+                                       if (peg$silentFails === 0) { peg$fail(peg$c79); }
+                                   }
+                               }
+                           } else {
+                               s5 = peg$FAILED;
+                           }
+                           if (s5 !== peg$FAILED) {
+                               s4 = [s4, s5];
+                               s3 = s4;
+                           } else {
+                               peg$currPos = s3;
+                               s3 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s3;
+                           s3 = peg$FAILED;
+                       }
+                       if (s3 === peg$FAILED) {
+                           s3 = peg$currPos;
+                           s4 = [];
+                           if (peg$c78.test(input.charAt(peg$currPos))) {
+                               s5 = input.charAt(peg$currPos);
+                               peg$currPos++;
+                           } else {
+                               s5 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c79); }
+                           }
+                           if (s5 !== peg$FAILED) {
+                               while (s5 !== peg$FAILED) {
+                                   s4.push(s5);
+                                   if (peg$c78.test(input.charAt(peg$currPos))) {
+                                       s5 = input.charAt(peg$currPos);
+                                       peg$currPos++;
+                                   } else {
+                                       s5 = peg$FAILED;
+                                       if (peg$silentFails === 0) { peg$fail(peg$c79); }
+                                   }
+                               }
+                           } else {
+                               s4 = peg$FAILED;
+                           }
+                           if (s4 !== peg$FAILED) {
+                               s5 = peg$currPos;
+                               if (input.charCodeAt(peg$currPos) === 46) {
+                                   s6 = peg$c76;
+                                   peg$currPos++;
+                               } else {
+                                   s6 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c77); }
+                               }
+                               if (s6 !== peg$FAILED) {
+                                   s7 = [];
+                                   if (peg$c78.test(input.charAt(peg$currPos))) {
+                                       s8 = input.charAt(peg$currPos);
+                                       peg$currPos++;
+                                   } else {
+                                       s8 = peg$FAILED;
+                                       if (peg$silentFails === 0) { peg$fail(peg$c79); }
+                                   }
+                                   while (s8 !== peg$FAILED) {
+                                       s7.push(s8);
+                                       if (peg$c78.test(input.charAt(peg$currPos))) {
+                                           s8 = input.charAt(peg$currPos);
+                                           peg$currPos++;
+                                       } else {
+                                           s8 = peg$FAILED;
+                                           if (peg$silentFails === 0) { peg$fail(peg$c79); }
+                                       }
+                                   }
+                                   if (s7 !== peg$FAILED) {
+                                       s6 = [s6, s7];
+                                       s5 = s6;
+                                   } else {
+                                       peg$currPos = s5;
+                                       s5 = peg$FAILED;
+                                   }
+                               } else {
+                                   peg$currPos = s5;
+                                   s5 = peg$FAILED;
+                               }
+                               if (s5 === peg$FAILED) {
+                                   s5 = null;
+                               }
+                               if (s5 !== peg$FAILED) {
+                                   s4 = [s4, s5];
+                                   s3 = s4;
+                               } else {
+                                   peg$currPos = s3;
+                                   s3 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s3;
+                               s3 = peg$FAILED;
+                           }
+                       }
+                       if (s3 !== peg$FAILED) {
+                           s2 = [s2, s3];
+                           s1 = s2;
+                       } else {
+                           peg$currPos = s1;
+                           s1 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s1;
+                       s1 = peg$FAILED;
+                   }
+                   if (s1 !== peg$FAILED) {
+                       peg$savedPos = s0;
+                       s1 = peg$c80(s1);
+                   }
+                   s0 = s1;
+                   peg$silentFails--;
+                   if (s0 === peg$FAILED) {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c73); }
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parsehtml_raw_string() {
+                   var s0, s1, s2, s3;
+   
+                   s0 = peg$currPos;
+                   if (input.charCodeAt(peg$currPos) === 60) {
+                       s1 = peg$c81;
+                       peg$currPos++;
+                   } else {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c82); }
+                   }
+                   if (s1 !== peg$FAILED) {
+                       s2 = [];
+                       s3 = peg$parsehtml_char();
+                       if (s3 === peg$FAILED) {
+                           s3 = peg$parsehtml_raw_string();
+                       }
+                       while (s3 !== peg$FAILED) {
+                           s2.push(s3);
+                           s3 = peg$parsehtml_char();
+                           if (s3 === peg$FAILED) {
+                               s3 = peg$parsehtml_raw_string();
+                           }
+                       }
+                       if (s2 !== peg$FAILED) {
+                           if (input.charCodeAt(peg$currPos) === 62) {
+                               s3 = peg$c83;
+                               peg$currPos++;
+                           } else {
+                               s3 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c84); }
+                           }
+                           if (s3 !== peg$FAILED) {
+                               peg$savedPos = s0;
+                               s1 = peg$c85(s2);
+                               s0 = s1;
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parsehtml_char() {
+                   var s0, s1, s2, s3, s4;
+   
+                   s0 = peg$currPos;
+                   s1 = [];
+                   s2 = peg$currPos;
+                   s3 = peg$currPos;
+                   peg$silentFails++;
+                   if (input.charCodeAt(peg$currPos) === 62) {
+                       s4 = peg$c83;
+                       peg$currPos++;
+                   } else {
+                       s4 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c84); }
+                   }
+                   if (s4 === peg$FAILED) {
+                       if (input.charCodeAt(peg$currPos) === 60) {
+                           s4 = peg$c81;
+                           peg$currPos++;
+                       } else {
+                           s4 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c82); }
+                       }
+                   }
+                   peg$silentFails--;
+                   if (s4 === peg$FAILED) {
+                       s3 = void 0;
+                   } else {
+                       peg$currPos = s3;
+                       s3 = peg$FAILED;
+                   }
+                   if (s3 !== peg$FAILED) {
+                       if (input.length > peg$currPos) {
+                           s4 = input.charAt(peg$currPos);
+                           peg$currPos++;
+                       } else {
+                           s4 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c86); }
+                       }
+                       if (s4 !== peg$FAILED) {
+                           peg$savedPos = s2;
+                           s3 = peg$c87(s4);
+                           s2 = s3;
+                       } else {
+                           peg$currPos = s2;
+                           s2 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s2;
+                       s2 = peg$FAILED;
+                   }
+                   if (s2 !== peg$FAILED) {
+                       while (s2 !== peg$FAILED) {
+                           s1.push(s2);
+                           s2 = peg$currPos;
+                           s3 = peg$currPos;
+                           peg$silentFails++;
+                           if (input.charCodeAt(peg$currPos) === 62) {
+                               s4 = peg$c83;
+                               peg$currPos++;
+                           } else {
+                               s4 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c84); }
+                           }
+                           if (s4 === peg$FAILED) {
+                               if (input.charCodeAt(peg$currPos) === 60) {
+                                   s4 = peg$c81;
+                                   peg$currPos++;
+                               } else {
+                                   s4 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c82); }
+                               }
+                           }
+                           peg$silentFails--;
+                           if (s4 === peg$FAILED) {
+                               s3 = void 0;
+                           } else {
+                               peg$currPos = s3;
+                               s3 = peg$FAILED;
+                           }
+                           if (s3 !== peg$FAILED) {
+                               if (input.length > peg$currPos) {
+                                   s4 = input.charAt(peg$currPos);
+                                   peg$currPos++;
+                               } else {
+                                   s4 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c86); }
+                               }
+                               if (s4 !== peg$FAILED) {
+                                   peg$savedPos = s2;
+                                   s3 = peg$c87(s4);
+                                   s2 = s3;
+                               } else {
+                                   peg$currPos = s2;
+                                   s2 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s2;
+                               s2 = peg$FAILED;
+                           }
+                       }
+                   } else {
+                       s1 = peg$FAILED;
+                   }
+                   if (s1 !== peg$FAILED) {
+                       peg$savedPos = s0;
+                       s1 = peg$c88(s1);
+                   }
+                   s0 = s1;
+   
+                   return s0;
+               }
+   
+               function peg$parseQUOTED_STRING() {
+                   var s0, s1, s2, s3, s4, s5, s6, s7;
+   
+                   s0 = peg$currPos;
+                   if (input.charCodeAt(peg$currPos) === 34) {
+                       s1 = peg$c89;
+                       peg$currPos++;
+                   } else {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c90); }
+                   }
+                   if (s1 !== peg$FAILED) {
+                       if (input.charCodeAt(peg$currPos) === 34) {
+                           s2 = peg$c89;
+                           peg$currPos++;
+                       } else {
+                           s2 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c90); }
+                       }
+                       if (s2 !== peg$FAILED) {
+                           peg$savedPos = s0;
+                           s1 = peg$c91();
+                           s0 = s1;
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+                   if (s0 === peg$FAILED) {
+                       s0 = peg$currPos;
+                       s1 = peg$currPos;
+                       if (input.charCodeAt(peg$currPos) === 34) {
+                           s2 = peg$c89;
+                           peg$currPos++;
+                       } else {
+                           s2 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c90); }
+                       }
+                       if (s2 !== peg$FAILED) {
+                           s3 = peg$parsechars();
+                           if (s3 !== peg$FAILED) {
+                               s4 = peg$currPos;
+                               if (input.charCodeAt(peg$currPos) === 92) {
+                                   s5 = peg$c92;
+                                   peg$currPos++;
+                               } else {
+                                   s5 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c93); }
+                               }
+                               if (s5 !== peg$FAILED) {
+                                   s6 = peg$parseNEWLINE();
+                                   if (s6 !== peg$FAILED) {
+                                       s7 = peg$parsechars();
+                                       if (s7 !== peg$FAILED) {
+                                           s5 = [s5, s6, s7];
+                                           s4 = s5;
+                                       } else {
+                                           peg$currPos = s4;
+                                           s4 = peg$FAILED;
+                                       }
+                                   } else {
+                                       peg$currPos = s4;
+                                       s4 = peg$FAILED;
+                                   }
+                               } else {
+                                   peg$currPos = s4;
+                                   s4 = peg$FAILED;
+                               }
+                               if (s4 === peg$FAILED) {
+                                   s4 = null;
+                               }
+                               if (s4 !== peg$FAILED) {
+                                   if (input.charCodeAt(peg$currPos) === 34) {
+                                       s5 = peg$c89;
+                                       peg$currPos++;
+                                   } else {
+                                       s5 = peg$FAILED;
+                                       if (peg$silentFails === 0) { peg$fail(peg$c90); }
+                                   }
+                                   if (s5 !== peg$FAILED) {
+                                       s2 = [s2, s3, s4, s5];
+                                       s1 = s2;
+                                   } else {
+                                       peg$currPos = s1;
+                                       s1 = peg$FAILED;
+                                   }
+                               } else {
+                                   peg$currPos = s1;
+                                   s1 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s1;
+                               s1 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s1;
+                           s1 = peg$FAILED;
+                       }
+                       if (s1 !== peg$FAILED) {
+                           s2 = peg$currPos;
+                           s3 = peg$parse_();
+                           if (s3 !== peg$FAILED) {
+                               if (input.charCodeAt(peg$currPos) === 43) {
+                                   s4 = peg$c94;
+                                   peg$currPos++;
+                               } else {
+                                   s4 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c95); }
+                               }
+                               if (s4 !== peg$FAILED) {
+                                   s5 = peg$parse_();
+                                   if (s5 !== peg$FAILED) {
+                                       s6 = peg$parseQUOTED_STRING();
+                                       if (s6 !== peg$FAILED) {
+                                           peg$savedPos = s2;
+                                           s3 = peg$c96(s1, s6);
+                                           s2 = s3;
+                                       } else {
+                                           peg$currPos = s2;
+                                           s2 = peg$FAILED;
+                                       }
+                                   } else {
+                                       peg$currPos = s2;
+                                       s2 = peg$FAILED;
+                                   }
+                               } else {
+                                   peg$currPos = s2;
+                                   s2 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s2;
+                               s2 = peg$FAILED;
+                           }
+                           if (s2 === peg$FAILED) {
+                               s2 = null;
+                           }
+                           if (s2 !== peg$FAILED) {
+                               peg$savedPos = s0;
+                               s1 = peg$c97(s1, s2);
+                               s0 = s1;
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parsechars() {
+                   var s0, s1, s2;
+   
+                   s0 = peg$currPos;
+                   s1 = [];
+                   s2 = peg$parsechar();
+                   if (s2 !== peg$FAILED) {
+                       while (s2 !== peg$FAILED) {
+                           s1.push(s2);
+                           s2 = peg$parsechar();
+                       }
+                   } else {
+                       s1 = peg$FAILED;
+                   }
+                   if (s1 !== peg$FAILED) {
+                       peg$savedPos = s0;
+                       s1 = peg$c98(s1);
+                   }
+                   s0 = s1;
+   
+                   return s0;
+               }
+   
+               function peg$parsechar() {
+                   var s0, s1, s2;
+   
+                   if (peg$c99.test(input.charAt(peg$currPos))) {
+                       s0 = input.charAt(peg$currPos);
+                       peg$currPos++;
+                   } else {
+                       s0 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c100); }
+                   }
+                   if (s0 === peg$FAILED) {
+                       s0 = peg$currPos;
+                       if (input.substr(peg$currPos, 2) === peg$c101) {
+                           s1 = peg$c101;
+                           peg$currPos += 2;
+                       } else {
+                           s1 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c102); }
+                       }
+                       if (s1 !== peg$FAILED) {
+                           peg$savedPos = s0;
+                           s1 = peg$c103();
+                       }
+                       s0 = s1;
+                       if (s0 === peg$FAILED) {
+                           s0 = peg$currPos;
+                           if (input.charCodeAt(peg$currPos) === 92) {
+                               s1 = peg$c92;
+                               peg$currPos++;
+                           } else {
+                               s1 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c93); }
+                           }
+                           if (s1 !== peg$FAILED) {
+                               s2 = peg$parseNEWLINE();
+                               if (s2 !== peg$FAILED) {
+                                   peg$savedPos = s0;
+                                   s1 = peg$c104();
+                                   s0 = s1;
+                               } else {
+                                   peg$currPos = s0;
+                                   s0 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                           if (s0 === peg$FAILED) {
+                               s0 = peg$currPos;
+                               if (input.charCodeAt(peg$currPos) === 92) {
+                                   s1 = peg$c92;
+                                   peg$currPos++;
+                               } else {
+                                   s1 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c93); }
+                               }
+                               if (s1 !== peg$FAILED) {
+                                   peg$savedPos = s0;
+                                   s1 = peg$c105();
+                               }
+                               s0 = s1;
+                           }
+                       }
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parseCOMMENT() {
+                   var s0, s1;
+   
+                   peg$silentFails++;
+                   s0 = peg$parseBLOCK_COMMENT();
+                   if (s0 === peg$FAILED) {
+                       s0 = peg$parseC_COMMENT();
+                       if (s0 === peg$FAILED) {
+                           s0 = peg$parseMACRO_COMMENT();
+                       }
+                   }
+                   peg$silentFails--;
+                   if (s0 === peg$FAILED) {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c106); }
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parseBLOCK_COMMENT() {
+                   var s0, s1, s2, s3, s4, s5;
+   
+                   peg$silentFails++;
+                   s0 = peg$currPos;
+                   if (input.substr(peg$currPos, 2) === peg$c108) {
+                       s1 = peg$c108;
+                       peg$currPos += 2;
+                   } else {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c109); }
+                   }
+                   if (s1 !== peg$FAILED) {
+                       s2 = [];
+                       s3 = peg$currPos;
+                       s4 = peg$currPos;
+                       peg$silentFails++;
+                       if (input.substr(peg$currPos, 2) === peg$c110) {
+                           s5 = peg$c110;
+                           peg$currPos += 2;
+                       } else {
+                           s5 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c111); }
+                       }
+                       peg$silentFails--;
+                       if (s5 === peg$FAILED) {
+                           s4 = void 0;
+                       } else {
+                           peg$currPos = s4;
+                           s4 = peg$FAILED;
+                       }
+                       if (s4 !== peg$FAILED) {
+                           if (input.length > peg$currPos) {
+                               s5 = input.charAt(peg$currPos);
+                               peg$currPos++;
+                           } else {
+                               s5 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c86); }
+                           }
+                           if (s5 !== peg$FAILED) {
+                               peg$savedPos = s3;
+                               s4 = peg$c112(s5);
+                               s3 = s4;
+                           } else {
+                               peg$currPos = s3;
+                               s3 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s3;
+                           s3 = peg$FAILED;
+                       }
+                       while (s3 !== peg$FAILED) {
+                           s2.push(s3);
+                           s3 = peg$currPos;
+                           s4 = peg$currPos;
+                           peg$silentFails++;
+                           if (input.substr(peg$currPos, 2) === peg$c110) {
+                               s5 = peg$c110;
+                               peg$currPos += 2;
+                           } else {
+                               s5 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c111); }
+                           }
+                           peg$silentFails--;
+                           if (s5 === peg$FAILED) {
+                               s4 = void 0;
+                           } else {
+                               peg$currPos = s4;
+                               s4 = peg$FAILED;
+                           }
+                           if (s4 !== peg$FAILED) {
+                               if (input.length > peg$currPos) {
+                                   s5 = input.charAt(peg$currPos);
+                                   peg$currPos++;
+                               } else {
+                                   s5 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c86); }
+                               }
+                               if (s5 !== peg$FAILED) {
+                                   peg$savedPos = s3;
+                                   s4 = peg$c112(s5);
+                                   s3 = s4;
+                               } else {
+                                   peg$currPos = s3;
+                                   s3 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s3;
+                               s3 = peg$FAILED;
+                           }
+                       }
+                       if (s2 !== peg$FAILED) {
+                           if (input.substr(peg$currPos, 2) === peg$c110) {
+                               s3 = peg$c110;
+                               peg$currPos += 2;
+                           } else {
+                               s3 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c111); }
+                           }
+                           if (s3 !== peg$FAILED) {
+                               peg$savedPos = s0;
+                               s1 = peg$c113(s2);
+                               s0 = s1;
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+                   peg$silentFails--;
+                   if (s0 === peg$FAILED) {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c107); }
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parseC_COMMENT() {
+                   var s0, s1, s2, s3, s4, s5;
+   
+                   peg$silentFails++;
+                   s0 = peg$currPos;
+                   if (input.substr(peg$currPos, 2) === peg$c115) {
+                       s1 = peg$c115;
+                       peg$currPos += 2;
+                   } else {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c116); }
+                   }
+                   if (s1 !== peg$FAILED) {
+                       s2 = [];
+                       s3 = peg$currPos;
+                       s4 = peg$currPos;
+                       peg$silentFails++;
+                       if (peg$c117.test(input.charAt(peg$currPos))) {
+                           s5 = input.charAt(peg$currPos);
+                           peg$currPos++;
+                       } else {
+                           s5 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c118); }
+                       }
+                       peg$silentFails--;
+                       if (s5 === peg$FAILED) {
+                           s4 = void 0;
+                       } else {
+                           peg$currPos = s4;
+                           s4 = peg$FAILED;
+                       }
+                       if (s4 !== peg$FAILED) {
+                           if (input.length > peg$currPos) {
+                               s5 = input.charAt(peg$currPos);
+                               peg$currPos++;
+                           } else {
+                               s5 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c86); }
+                           }
+                           if (s5 !== peg$FAILED) {
+                               peg$savedPos = s3;
+                               s4 = peg$c15(s5);
+                               s3 = s4;
+                           } else {
+                               peg$currPos = s3;
+                               s3 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s3;
+                           s3 = peg$FAILED;
+                       }
+                       while (s3 !== peg$FAILED) {
+                           s2.push(s3);
+                           s3 = peg$currPos;
+                           s4 = peg$currPos;
+                           peg$silentFails++;
+                           if (peg$c117.test(input.charAt(peg$currPos))) {
+                               s5 = input.charAt(peg$currPos);
+                               peg$currPos++;
+                           } else {
+                               s5 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c118); }
+                           }
+                           peg$silentFails--;
+                           if (s5 === peg$FAILED) {
+                               s4 = void 0;
+                           } else {
+                               peg$currPos = s4;
+                               s4 = peg$FAILED;
+                           }
+                           if (s4 !== peg$FAILED) {
+                               if (input.length > peg$currPos) {
+                                   s5 = input.charAt(peg$currPos);
+                                   peg$currPos++;
+                               } else {
+                                   s5 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c86); }
+                               }
+                               if (s5 !== peg$FAILED) {
+                                   peg$savedPos = s3;
+                                   s4 = peg$c15(s5);
+                                   s3 = s4;
+                               } else {
+                                   peg$currPos = s3;
+                                   s3 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s3;
+                               s3 = peg$FAILED;
+                           }
+                       }
+                       if (s2 !== peg$FAILED) {
+                           if (peg$c117.test(input.charAt(peg$currPos))) {
+                               s3 = input.charAt(peg$currPos);
+                               peg$currPos++;
+                           } else {
+                               s3 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c118); }
+                           }
+                           if (s3 === peg$FAILED) {
+                               s3 = null;
+                           }
+                           if (s3 !== peg$FAILED) {
+                               peg$savedPos = s0;
+                               s1 = peg$c119(s2);
+                               s0 = s1;
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+                   peg$silentFails--;
+                   if (s0 === peg$FAILED) {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c114); }
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parseMACRO_COMMENT() {
+                   var s0, s1, s2, s3, s4, s5;
+   
+                   peg$silentFails++;
+                   s0 = peg$currPos;
+                   if (input.charCodeAt(peg$currPos) === 35) {
+                       s1 = peg$c121;
+                       peg$currPos++;
+                   } else {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c122); }
+                   }
+                   if (s1 !== peg$FAILED) {
+                       s2 = [];
+                       s3 = peg$currPos;
+                       s4 = peg$currPos;
+                       peg$silentFails++;
+                       if (peg$c117.test(input.charAt(peg$currPos))) {
+                           s5 = input.charAt(peg$currPos);
+                           peg$currPos++;
+                       } else {
+                           s5 = peg$FAILED;
+                           if (peg$silentFails === 0) { peg$fail(peg$c118); }
+                       }
+                       peg$silentFails--;
+                       if (s5 === peg$FAILED) {
+                           s4 = void 0;
+                       } else {
+                           peg$currPos = s4;
+                           s4 = peg$FAILED;
+                       }
+                       if (s4 !== peg$FAILED) {
+                           if (input.length > peg$currPos) {
+                               s5 = input.charAt(peg$currPos);
+                               peg$currPos++;
+                           } else {
+                               s5 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c86); }
+                           }
+                           if (s5 !== peg$FAILED) {
+                               peg$savedPos = s3;
+                               s4 = peg$c15(s5);
+                               s3 = s4;
+                           } else {
+                               peg$currPos = s3;
+                               s3 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s3;
+                           s3 = peg$FAILED;
+                       }
+                       while (s3 !== peg$FAILED) {
+                           s2.push(s3);
+                           s3 = peg$currPos;
+                           s4 = peg$currPos;
+                           peg$silentFails++;
+                           if (peg$c117.test(input.charAt(peg$currPos))) {
+                               s5 = input.charAt(peg$currPos);
+                               peg$currPos++;
+                           } else {
+                               s5 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c118); }
+                           }
+                           peg$silentFails--;
+                           if (s5 === peg$FAILED) {
+                               s4 = void 0;
+                           } else {
+                               peg$currPos = s4;
+                               s4 = peg$FAILED;
+                           }
+                           if (s4 !== peg$FAILED) {
+                               if (input.length > peg$currPos) {
+                                   s5 = input.charAt(peg$currPos);
+                                   peg$currPos++;
+                               } else {
+                                   s5 = peg$FAILED;
+                                   if (peg$silentFails === 0) { peg$fail(peg$c86); }
+                               }
+                               if (s5 !== peg$FAILED) {
+                                   peg$savedPos = s3;
+                                   s4 = peg$c15(s5);
+                                   s3 = s4;
+                               } else {
+                                   peg$currPos = s3;
+                                   s3 = peg$FAILED;
+                               }
+                           } else {
+                               peg$currPos = s3;
+                               s3 = peg$FAILED;
+                           }
+                       }
+                       if (s2 !== peg$FAILED) {
+                           if (peg$c117.test(input.charAt(peg$currPos))) {
+                               s3 = input.charAt(peg$currPos);
+                               peg$currPos++;
+                           } else {
+                               s3 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c118); }
+                           }
+                           if (s3 === peg$FAILED) {
+                               s3 = null;
+                           }
+                           if (s3 !== peg$FAILED) {
+                               peg$savedPos = s0;
+                               s1 = peg$c119(s2);
+                               s0 = s1;
+                           } else {
+                               peg$currPos = s0;
+                               s0 = peg$FAILED;
+                           }
+                       } else {
+                           peg$currPos = s0;
+                           s0 = peg$FAILED;
+                       }
+                   } else {
+                       peg$currPos = s0;
+                       s0 = peg$FAILED;
+                   }
+                   peg$silentFails--;
+                   if (s0 === peg$FAILED) {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c120); }
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parse_() {
+                   var s0, s1;
+   
+                   peg$silentFails++;
+                   s0 = [];
+                   s1 = peg$parseWHITESPACE();
+                   if (s1 === peg$FAILED) {
+                       s1 = peg$parseCOMMENT();
+                   }
+                   while (s1 !== peg$FAILED) {
+                       s0.push(s1);
+                       s1 = peg$parseWHITESPACE();
+                       if (s1 === peg$FAILED) {
+                           s1 = peg$parseCOMMENT();
+                       }
+                   }
+                   peg$silentFails--;
+                   if (s0 === peg$FAILED) {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c123); }
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parseNEWLINE() {
+                   var s0, s1;
+   
+                   s0 = [];
+                   if (peg$c124.test(input.charAt(peg$currPos))) {
+                       s1 = input.charAt(peg$currPos);
+                       peg$currPos++;
+                   } else {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c125); }
+                   }
+                   if (s1 !== peg$FAILED) {
+                       while (s1 !== peg$FAILED) {
+                           s0.push(s1);
+                           if (peg$c124.test(input.charAt(peg$currPos))) {
+                               s1 = input.charAt(peg$currPos);
+                               peg$currPos++;
+                           } else {
+                               s1 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c125); }
+                           }
+                       }
+                   } else {
+                       s0 = peg$FAILED;
+                   }
+   
+                   return s0;
+               }
+   
+               function peg$parseWHITESPACE() {
+                   var s0, s1;
+   
+                   s0 = [];
+                   if (peg$c126.test(input.charAt(peg$currPos))) {
+                       s1 = input.charAt(peg$currPos);
+                       peg$currPos++;
+                   } else {
+                       s1 = peg$FAILED;
+                       if (peg$silentFails === 0) { peg$fail(peg$c127); }
+                   }
+                   if (s1 === peg$FAILED) {
+                       s1 = peg$parseNEWLINE();
+                   }
+                   if (s1 !== peg$FAILED) {
+                       while (s1 !== peg$FAILED) {
+                           s0.push(s1);
+                           if (peg$c126.test(input.charAt(peg$currPos))) {
+                               s1 = input.charAt(peg$currPos);
+                               peg$currPos++;
+                           } else {
+                               s1 = peg$FAILED;
+                               if (peg$silentFails === 0) { peg$fail(peg$c127); }
+                           }
+                           if (s1 === peg$FAILED) {
+                               s1 = peg$parseNEWLINE();
+                           }
+                       }
+                   } else {
+                       s0 = peg$FAILED;
+                   }
+   
+                   return s0;
+               }
+   
+               peg$result = peg$startRuleFunction();
+   
+               if (peg$result !== peg$FAILED && peg$currPos === input.length) {
+                   return peg$result;
+               } else {
+                   if (peg$result !== peg$FAILED && peg$currPos < input.length) {
+                       peg$fail(peg$endExpectation());
+                   }
+   
+                   throw peg$buildStructuredError(
+                       peg$maxFailExpected,
+                       peg$maxFailPos < input.length ? input.charAt(peg$maxFailPos) : null,
+                       peg$maxFailPos < input.length
+                           ? peg$computeLocation(peg$maxFailPos, peg$maxFailPos + 1)
+                           : peg$computeLocation(peg$maxFailPos, peg$maxFailPos)
+                   );
+               }
+           }
+   
+           return {
+               SyntaxError: peg$SyntaxError,
+               parse:       peg$parse
+           };
+       })();
+
