+++ src/view-browser.html
@@ -143,7 +143,6 @@
 <script type='text/javascript' src='cntk.js'></script>
 <script type='text/javascript' src='sklearn.js'></script>
 <script type='text/javascript' src='openvino-ir.js'></script>
-<script type='text/javascript' src='openvino-dot.js'></script>
 <script type='text/javascript' src='view-grapher.js'></script>
 <script type='text/javascript' src='view-sidebar.js'></script>
 <script type='text/javascript' src='view.js'></script>
