+++ src/view.js
@@ -19,7 +19,6 @@ var sklearn = sklearn || require('./sklearn');
 var tf = tf || require('./tf');
 var tflite = tflite || require('./tflite');
 var openvinoIR = openvinoIR || require('./openvino-ir');
-var openvinoDot = openvinoDot || require('./openvino-dot');
 
 var d3 = d3 || require('d3');
 var dagre = dagre || require('dagre');
@@ -970,8 +969,7 @@ view.ModelFactoryService = class {
             new tf.ModelFactory(),
             new sklearn.ModelFactory(),
             new cntk.ModelFactory(),
-            new openvinoIR.ModelFactory(),
-            new openvinoDot.ModelFactory()
+            new openvinoIR.ModelFactory()
         ];
     }
